Implement work add endpoint with database

+++ src/main/java/com/example/demo/product/CountableEfficiency.java
@@ -1,5 +1,7 @@
 package com.example.demo.product;
 
+import java.math.BigDecimal;
+
 public interface CountableEfficiency {
-    float countEfficiency();
+    BigDecimal countEfficiency();
 }
